requirements.txt
accelerate==1.13.0
APScheduler==3.11.2
yfinance>=0.2.40
peft==0.18.1
fastapi==0.122.0
uvicorn==0.38.0
playwright==1.58.0
playwright-stealth==2.0.3
pyarrow==23.0.1
schedule==1.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chronos-forecasting @ git+https://github.com/amazon-science/chronos-forecasting.git@6d68ed7c4ed2805d122d77b4660765b4089de5ca
click==8.3.2
contourpy==1.3.3
cycler==0.12.1
einops==0.8.2
et_xmlfile==2.0.0
fake-useragent==2.2.0
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==9.0.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mlx==0.31.1
mlx-lm==0.31.2
mlx-metal==0.31.1
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
ollama==0.6.1
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
pillow==12.2.0
protobuf==7.34.1
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==14.3.4
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
squarify==0.4.4
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==4.57.6
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
vnai==2.4.6
vnstock==3.5.1
vnstock3==3.2.1
vnstock_ezchart==0.0.3
wordcloud==1.9.6
zipp==3.23.0
